server/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
fastapi-users[sqlalchemy]
pydantic
python-jose[cryptography]
passlib
starlette